Antes de implementar punteros

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: tamamart <tamamart@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/19 21:27:19 by tamamart          #+#    #+#             */
-/*   Updated: 2025/10/21 18:06:54 by tamamart         ###   ########.fr       */
+/*   Updated: 2025/10/21 18:14:11 by tamamart         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int		ft_format(char character, va_list args);
 int		ft_putchar(char c);
 int		ft_putnbr(int c);
 int		ft_putstr(char *s);
-int		ft_putnbr_base(unsigned int nb, char *base, int len);
+int		ft_putnbr_base(unsigned long nb, char *base, int len);
 int		ft_strlen(char *str);
 
 #endif
